ExpFamilyDistributions package

ExpFamilyDistributions is a Julia package for using distributions member of the Exponential family. This package is mostly designed to build machine learning Bayesian models. For a general purpose statistical Julia package we recommend to use Distributions.jl.

Authors

Lucas Ondel, Brno University of technology, 2020.

Manual Outline

The package can be installed with the Julia package manager. From the Julia REPL, type ] to enter the Pkg REPL mode and run:

pkg> add ExpFamilyDistributions

Exponential Family super-type

All the distributions provided by this package are members of the exponential family of distribution, i.e. they have the follotwing canonical form:

\[p(x) = \exp \{ \eta^\top T(x) - A(\eta) + B(x) \}\]

where:

$\eta$ is the natural parameter (scalar or vector)

$T(x)$ is the sufficient statistic (scalar or vector)

$A(\eta)$ is the log-normalizer (scalar)

$B(x)$ is the base measure (scalar)

Practically, the package provide the following abstract type

ExpFamilyDistributions.ExpFamilyDistribution — Type

abstract type ExpFamilyDistribution end

Supertype for distributions member of the exponential family.

and each subtype of implements the following interface:

basemeasure

gradlognorm

kldiv

loglikelihood

lognorm

mean

naturalparam

stats

update!

ExpFamilyDistributions.basemeasure — Function

basemeasure(p, x)

Returns the base measure of x for the distribution p.

ExpFamilyDistributions.gradlognorm — Function

gradlognorm(p)

Returns the gradient of the log-normalizer of p w.r.t. its natural parameters.

ExpFamilyDistributions.kldiv — Function

kldiv(q::T, p::T) where T<:ExpFamilyDistribution

Compute the KL-divergence between two distributions of the same type (i.e. kldiv(Normal, Normal), kldiv(Dirichlet, Dirichlet), ...)

ExpFamilyDistributions.lognorm — Function

lognorm(p)

Returns the log-normalization constant of p.

ExpFamilyDistributions.loglikelihood — Function

loglikelihood(p, x)

Returns the log-likelihood of x for the distribution p.

ExpFamilyDistributions.mean — Function

mean(p)

Returns the mean of the distribution p.

ExpFamilyDistributions.naturalparam — Function

naturalparam(p)

Returns the natural parameters of p.

ExpFamilyDistributions.stats — Function

stats(p, x)

Returns the sufficient statistics of x for the distribution p.

ExpFamilyDistributions.update! — Function

update!(p, η)

Updates the parameters given a new natural parameter η.

Multivariate Normal distribution

Likelihood:

\[\mathcal{N}(\mu, \Sigma) = \frac{1}{(2\pi)^{\frac{D}{2}} | \Sigma |^{\frac{1}{2}}} \exp \big\{ -\frac{1}{2} (x - \mu)^\top \Sigma^{-1} (x - \mu) \big\}\]

Terms of the canonical form:

\[\begin{aligned} \eta &= \begin{bmatrix} \Sigma^{-1} \mu \\ -\frac{1}{2}\text{vec}(\Sigma^{-1}) \end{bmatrix} \\ T(x) &= \begin{bmatrix} x \\ \text{vec}(xx^\top) \end{bmatrix} \\ A(\eta) &= \frac{1}{2} \ln |\Sigma| + \frac{1}{2} \mu^\top \Sigma^{-1}\mu \\ B(x) &= -\frac{D}{2} \ln 2\pi \\ \nabla_{\eta} A(\eta) &= \begin{bmatrix} \mu \\ \Sigma + \mu \mu^\top \end{bmatrix} \end{aligned}\]

ExpFamilyDistributions.Normal — Type

mutable struct Normal{T,D} <: ExpFamilyDistribution
    μ
    Σ
end

Normal distribution with full covariance matrix.

Constructors

Normal{T,D}()

where T is the encoding type of the parameters and D is the dimension of the support.

Normal(μ[, Σ])

where μ is a vector and Σ is an symmetric matrix.

Examples

julia> Normal{Float32,2}()
Normal{Float32,2}:
  μ = Float32[0.0, 0.0]
  Σ = Float32[1.0 0.0; 0.0 1.0]

julia> Normal([1.0, 1.0])
Normal{Float64,2}:
  μ = [1.0, 1.0]
  Σ = [1.0 0.0; 0.0 1.0]

julia> using LinearAlgebra; Normal([1.0, 1.0], Symmetric([2.0 0.5; 0.5 1.0]))
Normal{Float64,2}:
  μ = [1.0, 1.0]
  Σ = [2.0 0.5; 0.5 1.0]

ExpFamilyDistributions.NormalDiag — Type

mutable struct NormalDiag{T,D} <: ExpFamilyDistribution
    μ
    v
end

Normal distribution with a diagonal covariance matrix. v is the diagonal of the covariance matrix. Note that you can still access the full covariance matrix by using the property Σ.

Constructors

NormalDiag{T,D}()

where T is the encoding type of the parameters and D is the dimension of the support.

NormalDiag(μ[, v])

where μ is a vector and v is the diagonal of the covariance matrix.

Examples

julia> NormalDiag{Float32, 2}()
NormalDiag{Float32,2}
  μ = Float32[0.0, 0.0]
  v = Float32[1.0, 1.0]

julia> NormalDiag([1.0, 1.0])
NormalDiag{Float64,2}
  μ = [1.0, 1.0]
  v = [1.0, 1.0]

julia> NormalDiag([1.0, 1.0], [2.0, 1.0])
NormalDiag{Float64,2}
  μ = [1.0, 1.0]
  v = [2.0, 1.0]

Gamma distribution

Likelihood:

\[\mathcal{G}(x | \alpha, \beta) = \frac{1}{\Gamma (\alpha)}\beta^{\alpha} x^{\alpha - 1} \exp \{ -\beta x \}\]

Terms of the canonical form:

\[\begin{aligned} \eta &= \begin{bmatrix} -\beta \\ \alpha \end{bmatrix} \\ T(x) &= \begin{bmatrix} x \\ \ln x \end{bmatrix} \\ A(\eta) &= \ln \Gamma(\alpha) - \alpha \ln \beta \\ B(x) &= -\ln x \\ \nabla_{\eta} A(\eta) &= \begin{bmatrix} \frac{\alpha}{\beta} \\ \psi(\alpha) - \ln\beta \end{bmatrix} \end{aligned}\]

ExpFamilyDistributions.Gamma — Type

mutable struct Gamma{T} <: ExpFamilyDistribution
    α
    β
end

Gamma distribution.

Constructors

Gamma{T}()

where T is the encoding type of the parameters.

Gamma{T}(α, β)

where T is a encoding type of the parameters and α and β are the parameters of the distribution.

Examples

julia> Gamma{Float32}()
Gamma{Float32}:
  α = 1.0
  β = 1.0

julia> Gamma{Float64}(1, 2)
Gamma{Float64}:
  α = 1.0
  β = 2.0

Dirichlet distribution

Likelihood:

\[\mathcal{D}(x | \alpha) = \frac{\Gamma(\sum_{i=1}^D \alpha_i)}{\prod_{i=1}^{D}\Gamma (\alpha_i)} \prod_{i=1}^D x_i^{\alpha - 1}\]

Terms of the canonical form:

\[\begin{aligned} \eta &= \alpha \\ T(x) &= \ln x \\ A(\eta) &= \sum_{i=1}^D \ln \Gamma(\alpha_i) - \ln \Gamma(\sum_{i=1}^D \alpha_i) \\ B(x) &= -\ln x \\ \nabla_{\eta} A(\eta) &= \begin{bmatrix} \psi(\alpha_1) - \psi(\sum_{i=1}^D \alpha_i) \\ \vdots \\ \psi(\alpha_D) - \psi(\sum_{i=1}^D \alpha_i) \\ \end{bmatrix} \end{aligned}\]

ExpFamilyDistributions.Dirichlet — Type

mutable struct Dirichlet{T,D} <: ExpFamilyDistribution
    α
end

Dirichlet distribution.

Constructors

Dirichlet{T,D}()

where T is the encoding type of the parameters and D is the dimension of the support.

Dirichlet(α)

where T is a encoding type of the parameters and α is a vector of parameters

Examples

julia> Dirichlet{Float32, 2}()
Dirichlet{Float32,2}:
  α = Float32[1.0, 1.0]

julia> Dirichlet([1.0, 2.0, 3.0])
Dirichlet{Float64,3}:
  α = [1.0, 2.0, 3.0]