δ-Distributions

Maximum likelihood (ML) and Maximum A Posteriori inference are special cases of Bayesian inference where the posterior is assumed to be a Dirac δ distribution:

\[\begin{aligned} \delta_{\mu}(x) &= \delta(x - \mu) = \begin{cases} \infty,& \text{if } x - \mu = 0 \\ 0 & \text{otherwise} \end{cases} \\ \int_x \delta_{\mu}(x) \text{d}x &= 1 \end{aligned}\]

To easily switch between Bayesian inference ML/MAP, the package provides the δ-distributions, i.e. Dirac δ dsistribution wrap around an "equivalent" distribution member of the exponential family.

ExpFamilyDistributions.δDistribution — Type

abstract type δDistribution

Supertype for the δ-distributions.

Each subtype of δDistribution implements partially the Exponential Family interface:

gradlognorm

mean

sample

splitgrad

update!.

δ-Normal distribution

ExpFamilyDistributions.δNormal — Type

mutable struct δNormal{T,D} <: δDistribution
    μ
end

The δ-equivalent of the Normal distribution.

Constructors

δNormal{T,D}()
δNormal(μ)

where T is the encoding type of the parameters and D is the dimension of the support and μ is the location of the Dirac δ pulse.

Examples

julia> δNormal{Float32,2}()
δNormal{Float32,2}:
  μ = Float32[0.0, 0.0]

julia> δNormal([1.0, 1.0])
δNormal{Float64,2}:
  μ = [1.0, 1.0]

ExpFamilyDistributions.δNormalDiag — Type

mutable struct δNormalDiag{T,D} <: δDistribution
    μ
end

The δ-equivalent of the NormalDiag distribution.

Constructors

δNormalDiag{T,D}()
δNormalDiag(μ)

where T is the encoding type of the parameters, D is the dimension of the support and μ is the location of the Dirac δ pulse.

Examples

julia> δNormalDiag{Float32, 2}()
δNormalDiag{Float32,2}:
  μ = Float32[0.0, 0.0]

julia> δNormalDiag([1.0, 1.0])
δNormalDiag{Float64,2}:
  μ = [1.0, 1.0]

δ-Gamma distribution

ExpFamilyDistributions.δGamma — Type

mutable struct δGamma{T} <: δDistribution
    μ
end

The δ-equivaltent of the Gamma distribution.

Constructors

δGamma{T}()
δGamma{T}(μ)

where T is the encoding type of the parameters and μ is the location of the Dirac δ pulse.

Examples

julia> δGamma{Float32}()
δGamma{Float32}:
  μ = 1.0

julia> δGamma{Float64}(2)
δGamma{Float64}:
  μ = 2.0

δ-Dirichlet distribution

ExpFamilyDistributions.δDirichlet — Type

mutable struct Dirichlet{T,D} <: δDistribution
    μ
end

The δ-equivalent of the Dirichlet distribution.

Constructors

Dirichlet{T,D}()
Dirichlet(μ)

where T is the encoding type of the parameters and D is the dimension of the support and μ is the location of the Dirac δ pulse.

Examples

julia> δDirichlet{Float32, 2}()
δDirichlet{Float32,2}:
  μ = Float32[0.5, 0.5]

julia> δDirichlet(Float32[1/2, 1/2])
δDirichlet{Float32,2}:
  μ = Float32[0.5, 0.5]

δ-Wishart distribution

ExpFamilyDistributions.δWishart — Type

mutable struct δWishart{T,D} <: δDistribution
    μ
end

The δ-equivaltent of the Wishart distribution.

Constructors

δWishart{T,D}()
δWishart(μ)

where T is the encoding type of the parameters and μ is the location of the Dirac δ pulse.

Examples

julia> δWishart{Float32,2}()
δWishart{Float32,2}:
  μ = Float32[1.0 0.0; 0.0 1.0]

julia> using LinearAlgebra; δWishart(Symmetric([1 0.5; 0.5 1]))
δWishart{Float64,2}:
  μ = [1.0 0.5; 0.5 1.0]